Perform Gibbs sampling for mediation MR without correlation

Model is that Y = (I+B)A + noise

Usage

gibbs_sampler(
  Y,
  Sd_hat,
  N,
  B,
  sigma,
  sigma1,
  sigma0,
  p,
  A,
  Z,
  alpha_B,
  beta_B,
  alpha_0,
  alpha_1,
  beta_0,
  beta_1,
  a,
  b
)

Arguments

Y

KxP arma::matrix of GWAS marginal associations

Sd_hat

KxP arma::matrix of GWAS std errors

N

number of samples to obtain

B

initial value of B arma::matrix

sigma

Initial value of prior std deviation for values of B

sigma1

K-arma::vector of Initial values of prior slab SDs for each trait's pleiotropy

sigma0

K-arma::vector of Initial values of prior splike SDs for each trait's pleiotropy

p

K-arma::vector of initial values for p, which is slab probability for mixture

A

KxP arma::matrix of initial values for A (ignored)

Z

KxP arma::matrix of initial values for Z (ignored), which is latent varaible indicating if A[i,j] comes from slab

alpha_B

Shape parameter of InvGamma prior for sigma (for B)

beta_B

Scale parameter of InvGamma prior for sigma (for B)

alpha_0

K-arma::vector Shape parameter of InvGamma prior for sigma0 (Spike)

alpha_1

K-arma::vector Shape parameter of InvGamma prior for sigma0 (Slab)

beta_0

K-arma::vector Scale parameter of InvGamma prior for sigma0 (Spike)

beta_1

K-arma::vector Scale parameter of InvGamma prior for sigma0 (Slab)

a

K-arma::vector First parameter of beta prior for p

b

K-arma::vector Second parameter of beta prior for p