Gibbs sampler

Given the observed effect and standard error, run gibbs sampler to draw samples from the posterior density

Usage

gibbs_wrapper(
  N,
  warmup,
  chains = 4,
  Gamma_hat,
  Sd_hat,
  sigma,
  sigma_1,
  sigma_0,
  p,
  parallel = TRUE,
  init = "Random",
  ratio = 10,
  cor_mat = NULL,
  Lambda = NULL,
  Lambda_inv = NULL
)

Arguments

N

Number of total iterations

warmup

Number of warm-up iterations.

chains

Number of Chains.

Gamma_hat

The observed effects

Sd_hat

The standard deviation of Gamma_hat

sigma

The prior mean for sigma

sigma_1

The prior mean for sigma_1

sigma_0

The prior mean for sigma_0

p

The prior mean for p

parallel

Indicating whether to parallelly run all the chains. Default is True.

init

Method of Initialization. Must be one of "EM" or "Random". Default is "Random"

ratio

The ratio of initial sigma1 and sigma0 when the initialization method is "Random". Default is 10.

cor_mat

The correlation matrix for noise. Default is null (no correlation)

Lambda

The block diagonal matrix such that each block is the covariance matrix of the noise for a SNP. Default is NULL (not provided)

Lambda_inv

The block diagonal matrix such that each block is the inverse of covariance matrix of the noise for a SNP. Default is NULL (not provided).

Value

A list of which the elements are outputs from each chain. Each of these outputs is a list with attributes

B

p

sigma1

sigma0

pi

sigma