Produce a summary table of the output from gibbs_wrapper.

Produce a summary table of the output from gibbs_wrapper.

Usage

summary_gibbs(result, pars, K, T = 1000, inv = FALSE)

Arguments

result

The output of gibbs_wrapper or gibbs_wrapper_cor

pars

A vector of parameters to be included in the summary table

K

The number of mediation layers

T

The warmup period. Default is 1000.

inv

When inv is TRUE, the function shows summary statistics for B in Gamma = B Gamma + alpha, otherwise it shows summary statistics for B in Gamma = (I + B) alpha. Default is False

Value

A table where each row corresponds to a parameter. It has columns:

mean

The average of samples after warm-up

variance

The variance of samples after warm-up

sd

The standard deviation of samples after warm-up

2.5%

The 2.5% quantile of samples after warm-up

50%

The 50

97.5%

The 97.5% quantile of samples after warm-up

ESS

The Effective sample sizes (combined from all the chains)

R_hat

A measure of convergence. A value below 1.1 indicates sign

of convergence