Perform EM Algorithm on Pleiotropy

Fit a two-component zero-mean Gaussian mixture on a sequence of observed values

Usage

zero.centered.em(
  values,
  std.errors,
  initial_slab_prob = 0.05,
  initial_sigma1_squared = 0.1,
  initial_sigma2_squared = 1e-04,
  iters = 100,
  optimizer = "Brent",
  show_outputs = FALSE
)

Arguments

values

A vector of observations.

initial_slab_prob

The initial value for the proportion of slab (the Gaussian component with higher variance). Default is .05.

initial_sigma1_squared

The initial value for the variance of the slab component. Default is .1.

initial_sigma2_squared

The initial value for the variance of the spike component. Default is .0001

iters

The number of iterations of EM algorithm. Default is 100

optimizer

The method used in Maximization Step. Must be one of "Brent", "Nelder-Mead" and "L-BFGS-B". Default is "Brent".

show_outputs

Whether to included a plot of the fitted density and the smoothed data density. Default is FALSE.

std.error

The standard error of noise of observations.

Value

A list of elements

S1

The standard deviation of the slab component

S2

The standard deviation of the spike component

Pi

The proportion of the slab component